Moved validator into qualities main & also added pylint: skip-file to __init__.py files
