+++ cSpell.json
@@ -100,6 +100,8 @@
     "BODMAS",
     "pypi",
     "matweb",
-    "behaviour"
+    "behaviour",
+    "pylintrc",
+    "isort"
   ]
 }
