+++ docs/readme.md
@@ -1,9 +1,28 @@
 ## Documentation
 
-### Hi, I am a placeholder made by william to be a placeholder
-The documentation will be done in .tex and than converted into a pdf.
+### Hi, I am a placeholder made by William to be a placeholder.
+
+The documentation will be written in LaTeX and compiled to PDF.
 
 > [!important]
-> An easy to understand python program is available here if you don't want to read.
+> An easy-to-understand Python program is available here if you don't want to read the full documentation.
 >
-> [introduction example](..\examples\introduction.uiv)
+> [Introduction example](..\examples\introduction.uiv)
+
+## Contributors
+
+Picounits is developed using:
+- `pylint` as the linter
+- CSpell for spell checking (Code Spell Checker, Bundled Dictionaries)
+
+> [!note]
+> Ensure that you add words to `picounits` cSpell.json, not your personal dictionary.
+
+My personal VS Code workflow also includes:
+- `Pylance`
+- `Python`
+- `Python Debugger`
+- `Python Environments`
+- `isort`
+- `Python Indent`
+- [UIV Language](https://github.com/wgbowley/uiv-language) package (supports `.ut` and `.uiv` formats)
