+++ src/picounits/blueprints/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/blueprints/__init__.py
