+++ src/picounits/configuration/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/configuration/__init__.py
