+++ src/picounits/core/__init__.py
@@ -1,6 +1,6 @@
+# pylint: skip-file
 # picounits/core/__init__.py
 
-from picounits.core.quantities.scalars.methods.validators import (
+from picounits.core.quantities.validators import (
     unit_validator, Quantity
 )
-
