+++ src/picounits/core/quantities/scalars/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/core/quantities/scalars/__init__.py
