+++ src/picounits/core/quantities/vectors/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/core/quantities/vectors/__init__.py
