+++ src/picounits/core/quantities/vectors/methods/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/core/quantities/vector/methods/__init__.py
