+++ src/picounits/core/quantities/vectors/types/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/core/quantities/vectors/types/__init__.py
