+++ src/picounits/extensions/__init__.py
@@ -1,3 +1,7 @@
+# pylint: skip-file
 # picounits/extensions/__init__.py
+
 from picounits.extensions.parser import Parser
 from picounits.extensions.loader import DynamicLoader
+
+_, _ = Parser, DynamicLoader
