+++ src/picounits/extensions/utilities/__init__.py
@@ -1,1 +1,2 @@
+# pylint: skip-file
 # picounits/extensions/utilities/__init__.py
