+++ src/picounits/py.typed
@@ -1,0 +1,1 @@
+# picounits/py.typed
