--- test.py
@@ -1,4 +1,0 @@
-from picounits import LENGTH, MICRO
-
-print(1 * MICRO * LENGTH ** 2)
-print((1 * MICRO * LENGTH) ** 2)
